requirements.txt
aiogram>=3.0
google-auth-oauthlib
google-auth
google-api-python-client
google-generativeai
python-dotenv
pydantic-settings
fastapi
uvicorn
